refactor: unify inventory click handling in FlipHelper and add utility method in GameUtils
